requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
freetype-py==2.2.0
glfw==2.0.0
glumpy==1.1.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
multipledispatch==0.6.0
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.5
opengles==0.0.1
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
pygame==2.0.1
Pygments==2.8.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrr==0.10.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
triangle==20200424
vispy==0.6.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
